Refactor ruby tests to use mocked command
Have refactored the ruby module tests to use a mocked command and no
longer depend on having a particular version of Ruby installed. While I
was here I also added a quick test for checking the formatting of the
ruby output.
