+++ src/modules/ruby.rs
@@ -45,3 +45,62 @@ fn format_ruby_version(ruby_version: &str) -> Option<String> {
     formatted_version.push_str(version);
     Some(formatted_version)
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::modules::utils::test::render_module;
+    use ansi_term::Color;
+    use std::fs::File;
+    use std::io;
+    use tempfile;
+
+    #[test]
+    fn folder_without_ruby_files() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+
+        let actual = render_module("ruby", dir.path());
+
+        let expected = None;
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+
+    #[test]
+    fn folder_with_gemfile() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        File::create(dir.path().join("Gemfile"))?.sync_all()?;
+
+        let actual = render_module("ruby", dir.path());
+
+        let expected = Some(format!("via {} ", Color::Red.bold().paint("💎 v2.5.1")));
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+
+    #[test]
+    fn folder_with_rb_file() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        File::create(dir.path().join("any.rb"))?.sync_all()?;
+
+        let actual = render_module("ruby", dir.path());
+
+        let expected = Some(format!("via {} ", Color::Red.bold().paint("💎 v2.5.1")));
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+
+    #[test]
+    fn test_format_ruby_version() -> io::Result<()> {
+        assert_eq!(
+            format_ruby_version("ruby 2.5.1p57 (2018-03-29 revision 63029) [x86_64-linux-gnu]"),
+            Some("v2.5.1".to_string())
+        );
+        assert_eq!(
+            format_ruby_version("ruby 2.7.0p0 (2019-12-25 revision 647ee6f091) [x86_64-linux-musl]"),
+            Some("v2.7.0".to_string())
+        );
+
+        Ok(())
+    }
+}
